Fixed the reduce_sum.
placeholders were defined as int8.  Chanign it to int32 fixed the issue. reduce_sum seems to retain the precision of the input. Since the sum is larger than 8 bit signed representation (max is 2^8 = 256), there is fixed point aliasing.  Infact a note on reduce_sum documentation notes: "Equivalent to np.sum apart the fact that numpy upcast uint8 and int32 to int64 while tensorflow returns the same dtype as the input."